--- PS17.c
@@ -1,1 +1,0 @@
-#include "PS17.h"
