--- keymaps/via/rules.mk
@@ -1,8 +1,0 @@
-VIA_ENABLE = yes
-LTO_ENABLE = yes
-LEADER_ENABLE = no
-FAUXCLICKY_ENABLE = no
-MIDI_ENABLE = no
-BLUETOOTH_ENABLE = no
-KEY_LOCK_ENABLE = no
-TERMINAL_ENABLE = no
